requirements.txt
groq==1.2.0
chromadb==0.5.3
posthog<3.0.0
sentence-transformers==3.0.1
tiktoken==0.7.0
gradio==4.42.0
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
googlemaps==4.10.0
serpapi==1.0.2
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4